+++ source/dotnet/Measurements.Abstractions/Api/Queries/GetByPeriodQuery.cs
@@ -1,0 +1,8 @@
+﻿using NodaTime;
+
+namespace Energinet.DataHub.Measurements.Abstractions.Api.Queries;
+
+/// <summary>
+/// Query for getting current measurements for a metering point for a period.
+/// </summary>
+public sealed record GetByPeriodQuery(string MeteringPointId, Instant From, Instant To);