+++ docs/Measurements.Client/ReleaseNotes/ReleaseNotes.md
@@ -1,5 +1,15 @@
 # Measurements.Client Release Notes
 
+## Version 7.1.1
+
+- fix link to documentation in package README
+- add `Measurements.Client` documentation
+
+## Version 7.0.0
+
+- Added two new dummy methods to the MeasurementsClient, `GetCurrentByPeriodAsync` and `GetAggregateByPeriodAsync`.
+- `Quantity` property of `MeasurementPointDto` is changed to a nullable `decimal`
+
 ## Version 6.0.0
 
 - Upgrade to dotnet version 9.0